requirements.txt
# Required Python packages for AWS-InfoBlox VPC Manager

pandas>=1.3.0
requests>=2.25.0
urllib3>=1.26.0
python-dotenv>=0.19.0
